+++ src/py_mono_tools/main.py
@@ -1,138 +1,28 @@
 """Contains all the commands that the CLI can execute."""
-import importlib
-import importlib.machinery
-import importlib.util
-import inspect
-import logging
 import os
 import os.path
 import pathlib
 import sys
 import typing as t
-from types import ModuleType
 
 import click
 
-from py_mono_tools.backends import Docker, System
 from py_mono_tools.config import consts, logger
-from py_mono_tools.goals import deployers as deployers_mod, linters as linters_mod, testers as testers_mod
-from py_mono_tools.goals.interface import Deployer, Language, Linter, Tester
+from py_mono_tools.goals.interface import Language
 from py_mono_tools.interface import GoalOutput
+from py_mono_tools.utils import (
+    filter_linters,
+    find_goals,
+    init_backend,
+    init_logger,
+    load_conf,
+    set_absolute_path,
+    set_path_from_conf_name,
+    set_relative_path,
+)
 
 
-def _find_goals():
-    goals = [
-        (linters_mod, Linter, "linters", consts.ALL_LINTERS, consts.ALL_LINTER_NAMES),
-        (deployers_mod, Deployer, "deployers", consts.ALL_DEPLOYERS, consts.ALL_DEPLOYER_NAMES),
-        (testers_mod, Tester, "testers", consts.ALL_TESTERS, consts.ALL_TESTER_NAMES),
-    ]
-
-    for goal in goals:
-        goal_mod, goal_abc, goal_name, consts_goal_instances, consts_goal_names = goal
-        logger.debug("Finding %s", goal_name)
-        goal_classes = inspect.getmembers(goal_mod, inspect.isclass)
-        goal_instances = [
-            goal_class[1]()
-            for goal_class in goal_classes
-            if issubclass(goal_class[1], goal_abc) and goal_class[1] != goal_abc
-        ]
-        consts_goal_instances.extend(goal_instances)
-
-        for goal_instance in goal_instances:
-            consts_goal_names.append(goal_instance.name)
-
-    consts.ALL_BACKENDS = [Docker, System]
-    consts.ALL_BACKEND_NAMES = [Docker.name, System.name]
-
-
-_find_goals()
-
-
-def _load_conf(conf_path: str) -> ModuleType:
-    conf_location = f"{conf_path}/CONF"
-    logger.debug("CONF location: %s", conf_location)
-
-    loader = importlib.machinery.SourceFileLoader("CONF", conf_location)
-    spec = importlib.util.spec_from_loader(loader.name, loader)
-    mod = importlib.util.module_from_spec(spec)  # type: ignore
-    loader.exec_module(mod)
-
-    return mod
-
-
-def _init_logger(verbose: bool, silent: bool = False):
-    logging_level = logging.INFO
-    if verbose:
-        logging_level = logging.DEBUG
-    if silent:
-        logging_level = logging.ERROR
-
-    logger.setLevel(logging_level)
-    for handler in logger.handlers:
-        handler.setLevel(logging_level)
-
-
-def _set_absolute_path(absolute_path: str):
-    logger.info("Overwriting execution root path: %s", absolute_path)
-
-    consts.EXECUTED_FROM = pathlib.Path(absolute_path).resolve()
-
-    os.chdir(consts.EXECUTED_FROM.resolve())
-
-
-def _set_relative_path(relative_path: str):
-    logger.info("Overwriting execution root path: %s", relative_path)
-
-    consts.EXECUTED_FROM = consts.EXECUTED_FROM.joinpath(pathlib.Path(relative_path).resolve())
-    os.chdir(consts.EXECUTED_FROM)
-
-
-def _set_path_from_conf_name(name: str):
-    logger.info("Setting path from conf name: %s", name)
-
-    for rel_path, _, filenames in os.walk("."):
-        for filename in filenames:
-            if filename == "CONF":
-                path = pathlib.Path(rel_path).resolve()
-                mod = _load_conf(str(path))
-                logger.debug("Found CONF: %s in %s", mod.NAME, path)
-                if mod.NAME.strip().lower() == name.strip().lower():
-                    logger.debug("Using CONF: %s in %s", mod.NAME, path)
-                    _set_absolute_path(str(path))
-                    return
-
-
-def _init_backend(_build_system: str):
-    logger.debug("Initializing build system: %s", _build_system)
-
-    consts.BACKENDS = {
-        Docker.name: Docker,
-        System.name: System,
-    }
-
-    consts.CURRENT_BACKEND = consts.BACKENDS[_build_system]()
-
-
-def _filter_linters(specific_linters: t.List[str], language: t.Optional[Language]) -> t.List[Linter]:
-    conf_linters: t.List[Linter] = consts.CONF.LINT  # type: ignore
-    logger.debug("Linters: %s", conf_linters)
-    linters_to_run: t.List[Linter] = []
-    specific_linters = [s.strip().lower() for s in specific_linters]
-
-    if language is None and not specific_linters:
-        logger.debug("No language or specific linters specified. Running all linters.")
-        linters_to_run = conf_linters
-    else:
-        logger.debug("Language: %s, specific: %s", language, specific_linters)
-        for linter in conf_linters:
-            if linter.name.strip().lower() in specific_linters:
-                linters_to_run.append(linter)
-                continue
-            if linter.language == language:
-                linters_to_run.append(linter)
-
-    logger.debug("Linters to run: %s", linters_to_run)
-    return linters_to_run
+find_goals()
 
 
 @click.group()
@@ -156,21 +46,21 @@ def cli(backend, absolute_path, relative_path, name, verbose, silent, machine_ou
     if "--help" in sys.argv or "-h" in sys.argv:
         return
 
-    _init_logger(verbose=verbose, silent=silent)
+    init_logger(verbose=verbose, silent=silent)
 
     logger.debug("Executed from: %s", consts.EXECUTED_FROM)
     logger.debug("Current backend: %s", consts.CURRENT_BACKEND)
     logger.debug("Backends: %s", consts.BACKENDS)
 
     if absolute_path is not None:
-        _set_absolute_path(absolute_path=absolute_path)
+        set_absolute_path(absolute_path=absolute_path)
     elif relative_path is not None:
-        _set_relative_path(relative_path=relative_path)
+        set_relative_path(relative_path=relative_path)
     elif name is not None:
-        _set_path_from_conf_name(name)
+        set_path_from_conf_name(name)
 
     try:
-        mod = _load_conf(consts.EXECUTED_FROM)
+        mod = load_conf(consts.EXECUTED_FROM)
         consts.CONF = mod
         if backend is None:
             try:
@@ -180,7 +70,7 @@ def cli(backend, absolute_path, relative_path, name, verbose, silent, machine_ou
 
         logger.info("Using backed: %s", backend)
 
-        _init_backend(backend)
+        init_backend(backend)
     except FileNotFoundError:
         logger.error("No CONF file found in %s", consts.EXECUTED_FROM)
 
@@ -241,7 +131,7 @@ def lint(
 
     logger.info("Starting lint")
 
-    linters_to_run = _filter_linters(specific_linters=specific, language=language)
+    linters_to_run = filter_linters(specific_linters=specific, language=language)
 
     if ignore_linter_weight is False:
         linters_to_run.sort(key=lambda x: x.weight, reverse=True)
@@ -314,7 +204,7 @@ def list_():
         for filename in filenames:
             if filename == "CONF":
                 path = pathlib.Path(rel_path).resolve()
-                mod = _load_conf(str(path))
+                mod = load_conf(str(path))
                 conf_names.append(f"{mod.NAME} -- {rel_path}")
 
     print("Backends:")
